Move accounts-db stats to a separate module
